Refactor FileMangerEditor's viewer
