+++ com.aobuchow.sample.commander/src/com/aobuchow/sample/commander/editor/FileManagerContentProvider.java
@@ -1,0 +1,75 @@
+package com.aobuchow.sample.commander.editor;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.stream.Collectors;
+
+import org.eclipse.core.resources.IFile;
+import org.eclipse.core.resources.IFolder;
+import org.eclipse.core.resources.IProject;
+import org.eclipse.core.resources.IResource;
+import org.eclipse.core.runtime.Adapters;
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.jface.viewers.IStructuredContentProvider;
+
+import com.aobuchow.sample.commander.resources.AudioFile;
+
+public class FileManagerContentProvider implements IStructuredContentProvider {
+
+	@Override
+	public Object[] getElements(Object inputElement) {
+		try {
+			return createModel((IResource) inputElement);
+		} catch (CoreException e) {
+			e.printStackTrace();
+		}
+
+		return null;
+	}
+
+	// TODO: Refactor this
+	private Object[] createModel(IResource o) throws CoreException {
+		Collection<Object> filesInSelectedDir = new ArrayList<Object>();
+		IFolder parentFolder = null;
+		if (!o.isAccessible() && !(o instanceof IProject)) {
+			return null;
+		}
+		if (o instanceof IProject) {
+			IProject project = Adapters.adapt(o, IProject.class);
+			filesInSelectedDir = Arrays.asList(project.members()).stream()
+					.map(FileManagerContentProvider::resourceConverter).collect(Collectors.toList());
+		} else if (o instanceof IFolder) {
+			parentFolder = Adapters.adapt(o, IFolder.class);
+
+			filesInSelectedDir = Arrays.asList(parentFolder.members()).stream()
+					.map(FileManagerContentProvider::resourceConverter).collect(Collectors.toList());
+
+		} else if (o instanceof IFile) {
+			parentFolder = Adapters.adapt(o.getParent(), IFolder.class);
+			if (parentFolder == null) {
+				IProject project = Adapters.adapt(o.getProject(), IProject.class);
+				filesInSelectedDir = Arrays.asList(project.members()).stream()
+						.map(FileManagerContentProvider::resourceConverter).collect(Collectors.toList());
+			} else {
+				filesInSelectedDir = Arrays.asList(parentFolder.members()).stream()
+						.map(FileManagerContentProvider::resourceConverter).collect(Collectors.toList());
+			}
+		}
+
+		if (parentFolder != null) {
+			filesInSelectedDir.add(parentFolder.getParent());
+		}
+
+		return filesInSelectedDir.stream().toArray(Object[]::new);
+	}
+
+	public static Object resourceConverter(IResource resource) {
+		if (resource.getType() == IResource.FOLDER) {
+			return Adapters.adapt(resource, IFolder.class);
+		} else {
+			return new AudioFile(resource);
+		}
+	}
+
+}
