+++ com.aobuchow.sample.commander/src/com/aobuchow/sample/commander/editor/FileManagerEditor.java
@@ -1,11 +1,6 @@
 package com.aobuchow.sample.commander.editor;
 
-import java.text.CharacterIterator;
-import java.text.StringCharacterIterator;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,10 +12,7 @@ import javax.annotation.PostConstruct;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.core.commands.operations.IOperationHistory;
 import org.eclipse.core.commands.operations.IUndoContext;
-import org.eclipse.core.filesystem.IFileStore;
 import org.eclipse.core.resources.IContainer;
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.resources.IResourceChangeEvent;
@@ -32,14 +24,9 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.e4.ui.di.Focus;
-import org.eclipse.jface.viewers.ArrayContentProvider;
-import org.eclipse.jface.viewers.ColumnLabelProvider;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.viewers.StructuredSelection;
-import org.eclipse.jface.viewers.TableViewer;
-import org.eclipse.jface.viewers.TableViewerColumn;
 import org.eclipse.jface.viewers.Viewer;
-import org.eclipse.jface.viewers.ViewerComparator;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.dnd.DND;
 import org.eclipse.swt.dnd.Transfer;
@@ -69,12 +56,10 @@ import org.eclipse.ui.part.ResourceTransfer;
 
 import com.aobuchow.sample.commander.Activator;
 import com.aobuchow.sample.commander.handlers.OpenResourceHandler;
-import com.aobuchow.sample.commander.resources.AudioFile;
 
 public class FileManagerEditor extends EditorPart implements IEditorPart {
 
-	private TableViewer viewer;
-	private Object[] model;
+	private FileManagerViewer viewer;
 	private IContainer inputContainer = null;
 	private IWorkspace workspace;
 	private IResourceChangeListener workspaceChangeListener;
@@ -85,67 +70,19 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 	@PostConstruct
 	@Override
 	public void createPartControl(Composite parent) {
-		// TODO: Extract the class to a viewer?
-		this.viewer = new TableViewer(parent, SWT.BORDER | SWT.FULL_SELECTION | SWT.MULTI);
-		this.viewer.setContentProvider(ArrayContentProvider.getInstance());
-		this.viewer.setComparator(new ResourceComparator());
-		TableViewerColumn nameColumn = createColumnFor(viewer, "File");
-		nameColumn.setLabelProvider(new ColumnLabelProvider() {
-			@Override
-			public String getText(Object element) {
-				if (element instanceof IResource) {
-					if ((element instanceof IFolder || element instanceof IProject)
-							&& element.equals(inputContainer.getParent())) {
-						return "..";
-					}
-					return ((IResource) element).getName();
-				}
-				return "null";
-			}
-
-			@Override
-			public Image getImage(Object element) {
-				if (element instanceof AudioFile) {
-					return Activator.imageDescriptorFromPlugin(Activator.PLUGIN_ID, "icons/AudioFile3_16.png")
-							.createImage();
-				}
-				if (element instanceof IContainer) {
-					return Activator.imageDescriptorFromPlugin(Activator.PLUGIN_ID, "icons/folder.png").createImage();
-				}
-				return null;
-			}
-		});
-		
-		TableViewerColumn sizeColumn = createColumnFor(viewer, "Size");
-		sizeColumn.setLabelProvider(new ColumnLabelProvider() {
-			@Override
-			public String getText(Object element) {
-
-				IFileStore fileStore;
-				try {
-					if (element instanceof IFile) {
-						fileStore = org.eclipse.core.filesystem.EFS.getStore(((IFile) element).getLocationURI());
-						return humanReadableByteCountSI(fileStore.fetchInfo().getLength());
-					}
-
-				} catch (CoreException e) {
-					e.printStackTrace();
-				}
-				return "";
-			}
-		});
-		
-		viewer.getControl().addKeyListener( new KeyListener() {
+		viewer = new FileManagerViewer(parent);
+		viewer.setInput(inputContainer);
+		viewer.getControl().addKeyListener(new KeyListener() {
 			@Override
 			public void keyReleased(KeyEvent e) {
-		
+
 			}
-			
+
 			@Override
 			public void keyPressed(KeyEvent e) {
 				// e.doit gets disabled regardless if successful as the beep it creates on Linux
 				// is annoying
-				switch(e.keyCode) {
+				switch (e.keyCode) {
 				case SWT.ARROW_RIGHT:
 					IStructuredSelection selection = viewer.getStructuredSelection();
 					if (selection.size() == 1 && selection.getFirstElement() instanceof IContainer) {
@@ -154,20 +91,11 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 					e.doit = false;
 					break;
 				case SWT.ARROW_LEFT:
-					if (!(inputContainer instanceof IProject)) {
-						changeActiveContainer(inputContainer.getParent());
+					if (!(viewer.getInput() instanceof IProject)) {
+						changeActiveContainer(((IContainer) viewer.getInput()).getParent());
 					}
 					e.doit = false;
 					break;
-				}
-			}
-		});
-
-		this.viewer.addSelectionChangedListener((event) -> {
-			if (event.getStructuredSelection().size() == 1) {
-				Object selection = event.getStructuredSelection().getFirstElement();
-				if (selection instanceof AudioFile) {
-					Activator.getDefault().getAudioPlayer().play((AudioFile) selection);
 				}
 			}
 		});
@@ -204,112 +132,8 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 		Transfer[] transferTypes = new Transfer[] { ResourceTransfer.getInstance() };
 		this.viewer.addDropSupport(operations, transferTypes, new FileManagerDropListener(this));
 		this.viewer.addDragSupport(operations, transferTypes, new FileManagerDragListener(viewer));
-		this.viewer.getTable().setLinesVisible(false);
-		this.viewer.getTable().setHeaderVisible(true);
-		this.viewer.setInput(model);
 
 		this.setPartName(inputContainer.getName());
-	}
-	
-	public static String humanReadableByteCountSI(long bytes) {
-	    if (-1000 < bytes && bytes < 1000) {
-	        return bytes + " B";
-	    }
-	    CharacterIterator ci = new StringCharacterIterator("kMGTPE");
-	    while (bytes <= -999_950 || bytes >= 999_950) {
-	        bytes /= 1000;
-	        ci.next();
-	    }
-	    return String.format("%.1f %cB", bytes / 1000.0, ci.current());
-	}
-
-	private TableViewerColumn createColumnFor(TableViewer viewer, String label) {
-		TableViewerColumn column = new TableViewerColumn(viewer, SWT.NONE);
-		column.getColumn().setWidth(400);
-		column.getColumn().setText(label);
-		column.getColumn().setMoveable(true);
-		return column;
-	}
-
-	public static Object resourceConverter(IResource resource) {
-		if (resource.getType() == IResource.FOLDER) {
-			return Adapters.adapt(resource, IFolder.class);
-		} else {
-			return new AudioFile(resource);
-		}
-	}
-
-	public class ResourceComparator extends ViewerComparator {
-		@Override
-		public int compare(Viewer viewer, Object e1, Object e2) {
-			// both are folders/projects, projects go first otherwise sort alphabetically
-			if (e1 instanceof IContainer && e2 instanceof IContainer) {
-				// the parent directory (..) should always be at the top
-				if (((IContainer) e1).equals(inputContainer.getParent())) {
-					return -1;
-				}
-				if (!e1.getClass().equals(e2.getClass())) {
-					if (e1 instanceof IProject) {
-						return -1;
-					} else {
-						return 1;
-					}
-				}
-				return Comparator.comparing(IContainer::getName, String.CASE_INSENSITIVE_ORDER).compare((IContainer) e1,
-						(IContainer) e2);
-			}
-			// one is audio file, one is folder/project, folder/project goes first
-			if (!e1.getClass().equals(e2.getClass())) {
-				if (e1 instanceof IContainer) {
-					return -1;
-				} else {
-					return 1;
-				}
-			}
-
-			// both are audio files, sort alphabetically
-			if (e1 instanceof AudioFile && e2 instanceof AudioFile) {
-				return Comparator.comparing(AudioFile::getName, String.CASE_INSENSITIVE_ORDER).compare((AudioFile) e1,
-						(AudioFile) e2);
-			}
-			return 0;
-		}
-	}
-
-	// TODO: Refactor this
-	private Object[] createModel(IResource o) throws CoreException {
-		Collection<Object> filesInSelectedDir = new ArrayList<Object>();
-		IFolder parentFolder = null;
-		if (!o.isAccessible() && !(o instanceof IProject)) {
-			return null;
-		}
-		if (o instanceof IProject) {
-			IProject project = Adapters.adapt(o, IProject.class);
-			filesInSelectedDir = Arrays.asList(project.members()).stream().map(FileManagerEditor::resourceConverter)
-					.collect(Collectors.toList());
-		} else if (o instanceof IFolder) {
-			parentFolder = Adapters.adapt(o, IFolder.class);
-
-			filesInSelectedDir = Arrays.asList(parentFolder.members()).stream()
-					.map(FileManagerEditor::resourceConverter).collect(Collectors.toList());
-
-		} else if (o instanceof IFile) {
-			parentFolder = Adapters.adapt(o.getParent(), IFolder.class);
-			if (parentFolder == null) {
-				IProject project = Adapters.adapt(o.getProject(), IProject.class);
-				filesInSelectedDir = Arrays.asList(project.members()).stream().map(FileManagerEditor::resourceConverter)
-						.collect(Collectors.toList());
-			} else {
-				filesInSelectedDir = Arrays.asList(parentFolder.members()).stream()
-						.map(FileManagerEditor::resourceConverter).collect(Collectors.toList());
-			}
-		}
-
-		if (parentFolder != null) {
-			filesInSelectedDir.add(parentFolder.getParent());
-		}
-
-		return filesInSelectedDir.stream().toArray(Object[]::new);
 	}
 
 	@Focus
@@ -326,7 +150,10 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 	public void dispose() {
 		viewer.getControl().dispose();
 		workspace.removeResourceChangeListener(workspaceChangeListener);
-		Activator.getDefault().getClipboard().dispose();
+		// Related to issue #31
+		// TODO: Only when the last FileManagerEditor closes should the clipboard be
+		// disposed
+		// Activator.getDefault().getClipboard().dispose();
 	}
 
 	@Override
@@ -386,16 +213,12 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 
 	@Override
 	public void init(IEditorSite site, IEditorInput input) throws PartInitException {
-		try {
-			if (input instanceof IContainerEditorInput) {
-				setInputFile(((IContainerEditorInput) input).getContainer());
-			} else if (input instanceof IFileEditorInput) {
-				setInputFile(((IFileEditorInput) input).getFile().getParent());
-			}
-			this.model = createModel(inputContainer);
-		} catch (CoreException e) {
-			e.printStackTrace();
+		if (input instanceof IContainerEditorInput) {
+			setInputFile(((IContainerEditorInput) input).getContainer());
+		} else if (input instanceof IFileEditorInput) {
+			setInputFile(((IFileEditorInput) input).getFile().getParent());
 		}
+
 		setSite(site);
 		// setInput is called in setInputFile
 		createPartControl(site.getShell());
@@ -404,14 +227,13 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 		workspaceChangeListener = event -> {
 			switch (event.getType()) {
 			case IResourceChangeEvent.POST_CHANGE:
-				this.refresh();
+				this.refreshViewerInput();
 				break;
 			}
 		};
 		workspace.addResourceChangeListener(workspaceChangeListener);
 
 		undoContext = (IUndoContext) ResourcesPlugin.getWorkspace().getAdapter(IUndoContext.class);
-
 		IWorkbench workbench = getSite().getWorkbenchWindow().getWorkbench();
 		history = workbench.getOperationSupport().getOperationHistory();
 
@@ -426,12 +248,15 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 	}
 
 	private void setInputFile(IContainer inputContainer) {
+		// should be
+		// this.model.setInpout(inputContainer)
 		this.inputContainer = inputContainer;
 		setInput(new ContainerEditorInput(inputContainer));
 		this.setTitleToolTip(inputContainer.getFullPath().toString());
 		firePropertyChange(IWorkbenchPartConstants.PROP_INPUT);
 		firePropertyChange(IWorkbenchPartConstants.PROP_TITLE);
-		// this.setPartName(inputContainer);
+		// this.setPartName(inputDirectory);
+		refreshViewerInput();
 	}
 
 	public void delete() {
@@ -440,29 +265,21 @@ public class FileManagerEditor extends EditorPart implements IEditorPart {
 		action.run();
 	}
 
-	// TODO: Always call this in setInputFile?
-	public void refresh() {
+	public void refreshViewerInput() {
 		if (this.viewer == null) {
 			return;
 		}
-		try {
-			this.model = createModel(inputContainer);
-			viewer.getControl().getDisplay().syncExec(() -> this.viewer.setInput(model));
-		} catch (CoreException e) {
-			e.printStackTrace();
-		}
+		viewer.getControl().getDisplay().syncExec(() -> this.viewer.setInput(inputContainer));
 	}
 
 	private void changeActiveContainer(IContainer container) {
 		IResource currentSelection = (IResource) this.viewer.getStructuredSelection().getFirstElement();
 		if (currentSelection != null) {
-			visitedDirectories.put(inputContainer, currentSelection);	
+			visitedDirectories.put(inputContainer, currentSelection);
 		}
 		setInputFile(container);
-		refresh();
 		IResource lastSelectedFile = visitedDirectories.get(container);
 		if (lastSelectedFile != null) {
-
 			try {
 				IResource[] model = (IResource[]) inputContainer.members();
 				Optional<IResource> newSelection = Arrays.asList(model).stream()
